Update nginx.conf
I added some of the important settings, but other things might be missing (e.g., SSL settings)  .  I think that specifying the version (which was here, but missing on genome) and more importantly, specifying the connection_upgrade, fixes this.   The IP stuff is important for all o fhte other redirection.